Implemented date filtering and UI improvements for the Records page
- Updated `activity_records.xml` with a new header, date picker button, and improved layout styling using the Digital-7 font.
- Added a date picker dialog to `Records.java` to allow filtering records by a specific day.
- Refactored `RecordsDao` and database logic to use `long` timestamps instead of `int`, ensuring correct Unix epoch handling.
- Renamed database to `food_database` and added `fallbackToDestructiveMigration()` to handle schema changes.
- Implemented bulk deletion functionality in `Records.java`.

+++ app/src/main/java/com/example/myapplication/RecordEntry.java
@@ -10,7 +10,7 @@ public class RecordEntry {
     @PrimaryKey(autoGenerate = true)
     public int uid;
     @ColumnInfo
-    public int date; //uses the unix timestamp
+    public long date; //uses the unix timestamp
     @ColumnInfo
     public String foodName;
     @ColumnInfo
